Read geopandas correctly
