+++ src/ireiat/data_pipeline/io_manager.py
@@ -17,7 +17,11 @@ def _get_read_function(format: str, metadata: dict = None) -> Callable:
     """Returns a function for file-reading based on the `format` and possibly `metadata`"""
     format_mapping = {
         "csv": pd.read_csv,
-        "parquet": pd.read_parquet,
+        "parquet": (
+            geopandas.read_parquet
+            if metadata and metadata.get("use_geopandas", False)
+            else pd.read_parquet
+        ),
         "zip": partial(pyogrio.read_dataframe, use_arrow=True),
         "txt": pd.read_table,
         "xlsx": pd.read_excel,
@@ -25,6 +29,7 @@ def _get_read_function(format: str, metadata: dict = None) -> Callable:
     if metadata and metadata.get("temp_unzip", False):
         return _temp_unzip_and_read
     if format in format_mapping:
+        print(format_mapping)
         return format_mapping[format]
     else:
         raise NotImplementedError(f"Cannot read file of type {format}!")
